+++ src/Storefront/Controller/GenesisReturnController.php
@@ -22,15 +22,16 @@ namespace Ecomprocessing\Genesis\Storefront\Controller;
 use Ecomprocessing\Genesis\Service\Flow\ReturnUrl;
 use Shopware\Storefront\Controller\StorefrontController;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\RedirectResponse;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
 class GenesisReturnController extends StorefrontController
 {
     /**
      * @var ReturnUrl
      */
-    private $returnUrlService;
+    private ReturnUrl $returnUrlService;
 
     public function __construct(ContainerInterface $container, ReturnUrl $return_url)
     {
@@ -39,16 +40,43 @@ class GenesisReturnController extends StorefrontController
     }
 
     /**
-     * @Route(
-     *     "/ecomprocessing/return/{token}",
-     *     name="frontend.ecomprocessing.return",
-     *     options={"seo"="false"},
-     *     methods={"GET"},
-     *     defaults={"auth_required"=false, "csrf_protected"=false, "_routeScope"={"storefront"}}
-     * )
+     * Redirect handler Controller action
      */
+    #[Route(
+        path:     '/ecomprocessing/return/{token}',
+        name:     'frontend.ecomprocessing.return',
+        options:  ['seo' => 'false'],
+        defaults: ['auth_required' => false, 'csrf_protected' => false, '_routeScope' => ['storefront']],
+        methods:  ['GET']
+    )]
     public function redirectOrderEndpoint(string $token): RedirectResponse
     {
         return $this->redirect($this->returnUrlService->getOrderEndpoint($token));
+    }
+
+    /**
+     * If the payment is processed into an iframe, response with a JS, breaking that iframe jail
+     *
+     * @param string $token
+     *
+     * @return Response
+     */
+    #[Route(
+        path:     '/ecomprocessing/return-iframe/{token}',
+        name:     'frontend.ecomprocessing.return_iframe',
+        options:  ['seo' => 'false'],
+        defaults: ['auth_required' => false, 'csrf_protected' => false, '_routeScope' => ['storefront']],
+        methods:  ['GET']
+    )]
+    public function redirectOrderEndpointBreakIframe(string $token): Response
+    {
+        $response = $this->render(
+            'storefront/iframehandler/index.html.twig',
+            ['returnUrl' => $this->returnUrlService->getOrderEndpoint($token)]
+        );
+        // In case we drop the support for 6.5
+        $response->headers->set('x-frame-options', 'ALLOWALL', true);
+
+        return $response;
     }
 }